[bluez-qt] Add (limited) support for a BlueZ 4 backend. Fixes JB#36054
Add limited support for a BlueZ 4 backend. The majority of the API is
supported; see the README for details on current support.

Instead of building kf5bluezqt packages, the spec sets
'Provides: kf5bluezqt' and builds either kf5bluezqt-bluez4-* or
kf5bluezqt-bluez5-* packages depending on the value of the
KF5BLUEZQT_BLUEZ_VERSION environment variable (defaulting to bluez5 if
this is not set).

BlueZ 4 support is #ifdef'd by the KF5BLUEZQT_BLUEZ_VERSION define.

Changes mainly consist of:

- Using the BlueZ 4 versions of the XML D-Bus interface definitions,
  added under bluezqt-/src/interfaces/bluez4
- Removing uses of DBusProperties
- Adding proxy agents so that the default agent can be reused as the
  application agent
- Returning PendingCall::NotSupported for features not provided by the
  BlueZ 4 API, such as MediaPlayer::play()
- Manually managing OBEX OPP sessions in OBEX Manager, as session
  support was limited to FTP sessions in BlueZ 4