[bluez-qt] Add Manager::pairWithDevice(QString) to pair with unknown devices. Contributes to JB#36678
In BlueZ 5, pairing is done via org.bluez.Device1.Pair(), and the
org.bluez.Device1 interface is persistently stored by the manager.

In BlueZ 4 however, devices come and go during discovery, so there may
not be an org.bluez.Device interface available for the device that you
want to pair with. This isn't an issue with BlueZ 4 as pairing is done
via org.bluez.Adapter.CreatePairedDevice() instead, but as the
kf5bluezqt API wraps the BlueZ 5 API, it's not possible to pair without
a DevicePtr.

Work around this by adding Manager::pairWithDevice() to pair with any
device, regardless of whether an org.bluez.Device interface is currently
available for that device with the BlueZ 4 backend. For BlueZ 5, this
will end up calling Device::pair() anyway, and for BlueZ 4, this will
call org.bluez.Adapter.CreatePairedDevice().