[bluez-qt] Check for object validity in macros
If an application connects to both the statusChanged and
transferredChanged signals for a transfer, and removes its reference to
the Transfer object when statusChanged=Complete, the application
crashes as ObexTransferPrivate::propertiesChanged() attempts to emit
the transferredChanged signal immediately after the statusChanged
signal, and q.data() is no longer valid. To avoid this, check for the
validity of q.data() in macros.h. For completeness, update
macros_bluez4.h to check for the validity of obj.