[bluez-qt] Don't connect to signals with QVariantMapMap parameters. Fixes JB#36800
Signals with QVariantMapMap parameters (i.e. ManagerPrivate::interfacesAdded
and ObexManagerPrivate::interfacesAdded) are not received even when
dbus-monitor indicates the signals are being sent. If the
QVariantMapMap parameter is removed, the signal is received as
expected.

This behaviour is application-dependent so I'm wondering if it is a
resurfacing of QTBUG-11485 and is dependent on the libraries used by
the application.

With this patch, the managers connect to interfacesAdded() without the
QVariantMapMap parameter, and when the signal is received, calls
DBusObjectManager::GetManagedObjects() to fetch the full list of
available interfaces, then updates the internal properties as per the
normal interfacesAdded() behaviour.