[bluez-qt] BlueZ 4: emit transferredChanged() for received files. Contributes to JB#36929
The 'Transferred' property was named 'Progress' in BlueZ 4 for
incoming files, so internally we must update the 'Transferred'
property so that transferredChanged() will be emitted when
ObexTransfer::propertiesChanged() is called. Otherwise, the
'Progress' value is set but does not trigger any property
changes as this is not a known value in propertiesChanged().

This should have been part of the JB#36929 change when the
functionality was moved into obextransfer_bluez4.cpp.