[bluez-qt] BlueZ 4: initialize Manager even if adapters are not yet available. Contributes to JB#36745
If Bluetooth is blocked, the manager initialization may fail and be
unable to find any adapters, depending on the Bluetooth/rfkill
interaction on the platform. However, ideally the manager can be
initialized even when adapters are not yet available, so that library
users do not have to check for when Bluetooth is unblocked etc. before
trying to initialize the manager.

However, this means that functionality that relies on the availability
of adapters may fail unexpectedly. For example, requestDefaultAgent()
will fail without an adapter in BlueZ 4, as internally the agent is
registered via org.bluez.Adapter, unlike in BlueZ 5, where the agent
is registered via org.bluez.AgentManager1 and does not require the
existence of an adapter.